Runs express server alongside electron app
Consolidates express server into server.js, updates bash and package scripts
